small changes to try and get a method calculate best learning factor in backpropogation to get working. not quite there yet, but small coommits are best practice, even if I dont always do it
