--- Release/nsightbuilddata
@@ -1,1 +1,0 @@
-Local@Native
