--- Release/sources.mk
@@ -1,29 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-CU_SRCS := 
-C_UPPER_SRCS := 
-CXX_SRCS := 
-C++_SRCS := 
-OBJ_SRCS := 
-CC_SRCS := 
-ASM_SRCS := 
-CPP_SRCS := 
-C_SRCS := 
-O_SRCS := 
-S_UPPER_SRCS := 
-CC_DEPS := 
-C++_DEPS := 
-EXECUTABLES := 
-C_UPPER_DEPS := 
-CXX_DEPS := 
-OBJS := 
-CU_DEPS := 
-CPP_DEPS := 
-C_DEPS := 
-
-# Every subdirectory with source files must be described here
-SUBDIRS := \
-. \
-
