--- Release/subdir.mk
@@ -1,28 +1,0 @@
-################################################################################
-# Automatically-generated file. Do not edit!
-################################################################################
-
-# Add inputs and outputs from these tool invocations to the build variables 
-CU_SRCS += \
-../cuNet.cu \
-../testAll.cu 
-
-OBJS += \
-./cuNet.o \
-./testAll.o 
-
-CU_DEPS += \
-./cuNet.d \
-./testAll.d 
-
-
-# Each subdirectory must supply rules for building sources it contributes
-%.o: ../%.cu
-	@echo 'Building file: $<'
-	@echo 'Invoking: NVCC Compiler'
-	/usr/local/cuda-9.2/bin/nvcc -O3 -gencode arch=compute_61,code=sm_61  -odir "." -M -o "$(@:%.o=%.d)" "$<"
-	/usr/local/cuda-9.2/bin/nvcc -O3 --compile --relocatable-device-code=false -gencode arch=compute_61,code=compute_61 -gencode arch=compute_61,code=sm_61  -x cu -o  "$@" "$<"
-	@echo 'Finished building: $<'
-	@echo ' '
-
-
