--- Release/testAll.d
@@ -1,137 +1,0 @@
-./testAll.o : ../testAll.cu \
-    /usr/include/stdc-predef.h \
-    /usr/local/cuda-9.2/bin/..//include/cuda_runtime.h \
-    /usr/local/cuda-9.2/bin/..//include/host_config.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/host_config.h \
-    /usr/include/features.h \
-    /usr/include/x86_64-linux-gnu/sys/cdefs.h \
-    /usr/include/x86_64-linux-gnu/bits/wordsize.h \
-    /usr/include/x86_64-linux-gnu/gnu/stubs.h \
-    /usr/include/x86_64-linux-gnu/gnu/stubs-64.h \
-    /usr/local/cuda-9.2/bin/..//include/builtin_types.h \
-    /usr/local/cuda-9.2/bin/..//include/device_types.h \
-    /usr/local/cuda-9.2/bin/..//include/host_defines.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/host_defines.h \
-    /usr/local/cuda-9.2/bin/..//include/driver_types.h \
-    /usr/local/cuda-9.2/bin/..//include/vector_types.h \
-    /usr/lib/gcc/x86_64-linux-gnu/5/include-fixed/limits.h \
-    /usr/lib/gcc/x86_64-linux-gnu/5/include-fixed/syslimits.h \
-    /usr/include/limits.h \
-    /usr/include/x86_64-linux-gnu/bits/posix1_lim.h \
-    /usr/include/x86_64-linux-gnu/bits/local_lim.h \
-    /usr/include/linux/limits.h \
-    /usr/include/x86_64-linux-gnu/bits/posix2_lim.h \
-    /usr/include/x86_64-linux-gnu/bits/xopen_lim.h \
-    /usr/include/x86_64-linux-gnu/bits/stdio_lim.h \
-    /usr/lib/gcc/x86_64-linux-gnu/5/include/stddef.h \
-    /usr/local/cuda-9.2/bin/..//include/surface_types.h \
-    /usr/local/cuda-9.2/bin/..//include/texture_types.h \
-    /usr/local/cuda-9.2/bin/..//include/library_types.h \
-    /usr/local/cuda-9.2/bin/..//include/channel_descriptor.h \
-    /usr/local/cuda-9.2/bin/..//include/cuda_runtime_api.h \
-    /usr/local/cuda-9.2/bin/..//include/cuda_device_runtime_api.h \
-    /usr/local/cuda-9.2/bin/..//include/driver_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/vector_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/vector_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/common_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/common_functions.h \
-    /usr/include/string.h \
-    /usr/include/xlocale.h \
-    /usr/include/x86_64-linux-gnu/bits/string3.h \
-    /usr/include/time.h \
-    /usr/include/x86_64-linux-gnu/bits/time.h \
-    /usr/include/x86_64-linux-gnu/bits/types.h \
-    /usr/include/x86_64-linux-gnu/bits/typesizes.h \
-    /usr/include/x86_64-linux-gnu/bits/timex.h \
-    /usr/include/c++/5/new \
-    /usr/include/x86_64-linux-gnu/c++/5/bits/c++config.h \
-    /usr/include/x86_64-linux-gnu/c++/5/bits/os_defines.h \
-    /usr/include/x86_64-linux-gnu/c++/5/bits/cpu_defines.h \
-    /usr/include/c++/5/exception \
-    /usr/include/c++/5/bits/atomic_lockfree_defines.h \
-    /usr/include/stdio.h \
-    /usr/include/libio.h \
-    /usr/include/_G_config.h \
-    /usr/include/wchar.h \
-    /usr/lib/gcc/x86_64-linux-gnu/5/include/stdarg.h \
-    /usr/include/x86_64-linux-gnu/bits/sys_errlist.h \
-    /usr/include/x86_64-linux-gnu/bits/stdio.h \
-    /usr/include/x86_64-linux-gnu/bits/stdio2.h \
-    /usr/include/stdlib.h \
-    /usr/include/x86_64-linux-gnu/bits/waitflags.h \
-    /usr/include/x86_64-linux-gnu/bits/waitstatus.h \
-    /usr/include/endian.h \
-    /usr/include/x86_64-linux-gnu/bits/endian.h \
-    /usr/include/x86_64-linux-gnu/bits/byteswap.h \
-    /usr/include/x86_64-linux-gnu/bits/byteswap-16.h \
-    /usr/include/x86_64-linux-gnu/sys/types.h \
-    /usr/include/x86_64-linux-gnu/sys/select.h \
-    /usr/include/x86_64-linux-gnu/bits/select.h \
-    /usr/include/x86_64-linux-gnu/bits/sigset.h \
-    /usr/include/x86_64-linux-gnu/bits/select2.h \
-    /usr/include/x86_64-linux-gnu/sys/sysmacros.h \
-    /usr/include/x86_64-linux-gnu/bits/pthreadtypes.h \
-    /usr/include/alloca.h \
-    /usr/include/x86_64-linux-gnu/bits/stdlib-bsearch.h \
-    /usr/include/x86_64-linux-gnu/bits/stdlib-float.h \
-    /usr/include/x86_64-linux-gnu/bits/stdlib.h \
-    /usr/include/assert.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/math_functions.h \
-    /usr/include/math.h \
-    /usr/include/x86_64-linux-gnu/bits/math-vector.h \
-    /usr/include/x86_64-linux-gnu/bits/libm-simd-decl-stubs.h \
-    /usr/include/x86_64-linux-gnu/bits/huge_val.h \
-    /usr/include/x86_64-linux-gnu/bits/huge_valf.h \
-    /usr/include/x86_64-linux-gnu/bits/huge_vall.h \
-    /usr/include/x86_64-linux-gnu/bits/inf.h \
-    /usr/include/x86_64-linux-gnu/bits/nan.h \
-    /usr/include/x86_64-linux-gnu/bits/mathdef.h \
-    /usr/include/x86_64-linux-gnu/bits/mathcalls.h \
-    /usr/include/x86_64-linux-gnu/bits/mathinline.h \
-    /usr/include/c++/5/cmath \
-    /usr/include/c++/5/bits/cpp_type_traits.h \
-    /usr/include/c++/5/ext/type_traits.h \
-    /usr/include/c++/5/cstdlib \
-    /usr/local/cuda-9.2/bin/..//include/crt/math_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/cuda_surface_types.h \
-    /usr/local/cuda-9.2/bin/..//include/cuda_texture_types.h \
-    /usr/local/cuda-9.2/bin/..//include/device_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/device_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/device_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/device_atomic_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/device_atomic_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/crt/device_double_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/device_double_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_20_atomic_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_20_atomic_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_32_atomic_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_32_atomic_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_35_atomic_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_60_atomic_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_60_atomic_functions.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_20_intrinsics.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_20_intrinsics.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_30_intrinsics.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_30_intrinsics.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_32_intrinsics.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_32_intrinsics.hpp \
-    /usr/local/cuda-9.2/bin/..//include/sm_35_intrinsics.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_61_intrinsics.h \
-    /usr/local/cuda-9.2/bin/..//include/sm_61_intrinsics.hpp \
-    /usr/local/cuda-9.2/bin/..//include/crt/sm_70_rt.h \
-    /usr/local/cuda-9.2/bin/..//include/crt/sm_70_rt.hpp \
-    /usr/local/cuda-9.2/bin/..//include/surface_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/texture_fetch_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/texture_indirect_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/surface_indirect_functions.h \
-    /usr/local/cuda-9.2/bin/..//include/device_launch_parameters.h \
-    ../testing_linear_algebra.h \
-    ../linear_algebra.h \
-    /usr/local/cuda-9.2/bin/..//include/cuda.h \
-    /usr/lib/gcc/x86_64-linux-gnu/5/include/stdint.h \
-    /usr/include/stdint.h \
-    /usr/include/x86_64-linux-gnu/bits/wchar.h \
-    ../testing_network.h \
-    ../network.h \
-    ../backpropogation.h \
-    ../database.h
