+++ cuNet/README.md
@@ -1,0 +1,19 @@
+# cuNet
+cuda based neural network
+
+
+database files contain the data used to train the networks as well as data managment. (I have not yet fully implemented file IO)
+
+linear_algebra file contains matrix and vector algebras relevant to simple neural networks
+
+network file contain matrix and vector data of weights and biases.
+
+backpropogate file contains the functions used to train a network.
+
+
+testAll.cu runs test of all functions.
+
+I run this code in ubuntu 16.04 LTS using eclipse Nsight.
+To do this I import the project from git using the project explorer.
+The compilarion of this project requires specific complilation settings detailed here.
+https://bit.ly/2wA43gl
