+++ cuNet/cuNet.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30225.117
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "cuNet", "cuNet.vcxproj", "{2E4D1CA2-318D-429E-8BD4-6A5C225B7BE3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Release|x64 = Release|x64
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2E4D1CA2-318D-429E-8BD4-6A5C225B7BE3}.Debug|x64.ActiveCfg = Debug|x64
+		{2E4D1CA2-318D-429E-8BD4-6A5C225B7BE3}.Debug|x64.Build.0 = Debug|x64
+		{2E4D1CA2-318D-429E-8BD4-6A5C225B7BE3}.Release|x64.ActiveCfg = Release|x64
+		{2E4D1CA2-318D-429E-8BD4-6A5C225B7BE3}.Release|x64.Build.0 = Release|x64
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {82D1B07A-7D14-453B-9DAC-61F16BC31D20}
+	EndGlobalSection
+EndGlobal
